fix(edge): Correct logic placement about enriching authData with user
Remove the getAuthData method
