+++ packages/edge/src/vercel-edge/utils/index.ts
@@ -1,2 +1,1 @@
-export * from './getAuthData';
 export * from './injectAuthIntoRequest';
