+++ packages/edge/src/vercel-edge/utils/injectAuthIntoRequest.ts
@@ -6,11 +6,11 @@ export function injectAuthIntoRequest(
   req: NextRequest,
   authData: AuthData,
 ): RequestWithAuth {
-  const { user, session, userId, sessionId } = authData;
+  const { user, session, userId, sessionId, getToken } = authData;
   const auth = {
     userId,
     sessionId,
-    getToken: authData.getToken,
+    getToken,
   };
 
   /* Object.assign is used here as NextRequest properties also include Symbols */
